initial restructure into interface structure
this still works for XML documents
